+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/settings/base.py
@@ -284,14 +284,12 @@ CRISPY_TEMPLATE_PACK = 'bootstrap4'
 DEFAULT_FROM_EMAIL = "{{cookiecutter.project_title}} <info@{{ cookiecutter.live_hostname }}>"
 SERVER_EMAIL = "{{cookiecutter.project_title}} server <server@{{ cookiecutter.live_hostname }}>"
 
-# Show emails in the console, but don't send them.
-EMAIL_BACKEND = 'django.core.mail.backends.console.EmailBackend'
+EMAIL_BACKEND = 'django.core.mail.backends.smtp.EmailBackend'
 
-# SMTP  --> This is only used in staging and production
-EMAIL_HOST = env.str('DJANGO_EMAIL_HOST', default='smtp.sparkpostmail.com')
-EMAIL_PORT = env.int('DJANGO_EMAIL_PORT', default=587)
-EMAIL_HOST_USER = env.str('DJANGO_EMAIL_HOST_USER', default='SMTP_Injection')
-EMAIL_HOST_PASSWORD = env.str('DJANGO_EMAIL_HOST_PASSWORD', default='')
+EMAIL_HOST = 'mailhog'
+EMAIL_PORT = 1025
+EMAIL_HOST_USER = ''
+EMAIL_HOST_PASSWORD = ''
 
 
 # Base logging config. Logs INFO and higher-level messages to console. Production-specific additions are in