+++ {{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/settings/production.py
@@ -7,6 +7,9 @@ ALLOWED_HOSTS = env.list('DJANGO_ALLOWED_HOSTS', default=['{{ cookiecutter.live_
 # Static site url, used when we need absolute url but lack request object, e.g. in email sending.
 SITE_URL = env.str('DJANGO_SITE_URL', default='https://{{ cookiecutter.live_hostname }}')
 
+EMAIL_HOST = env.str('DJANGO_EMAIL_HOST', default='smtp.sparkpostmail.com')
+EMAIL_PORT = env.int('DJANGO_EMAIL_PORT', default=587)
+EMAIL_HOST_USER = env.str('DJANGO_EMAIL_HOST_USER', default='TODO')
 EMAIL_HOST_PASSWORD = env.str('DJANGO_EMAIL_HOST_PASSWORD', default='TODO (api key)')
 
 RAVEN_BACKEND_DSN = env.str('DJANGO_RAVEN_BACKEND_DSN', default='https://TODO:TODO@sentry.thorgate.eu/TODO')